full replacement of panelapp
