--- test/input/panel_changes_expected.json
@@ -1,29 +1,0 @@
-{
- "metadata": {
-  "version": "0.11088"
- },
- "ENSG00ABCD": {
-  "symbol": "ABCD",
-  "moi": "BIALLELIC",
-  "new": false,
-  "changed": true,
-  "old_moi": "MONOALLELIC",
-  "flags": []
- },
- "ENSG00EFGH": {
-  "symbol": "EFGH",
-  "moi": "MONOALLELIC",
-  "new": true,
-  "changed": false,
-  "old_moi": null,
-  "flags": []
- },
- "ENSG00IJKL": {
-  "symbol": "IJKL",
-  "moi": "BOTH",
-  "new": false,
-  "changed": false,
-  "old_moi": null,
-  "flags": []
- }
-}
