+++ test/test_hail_categories.py
@@ -238,26 +238,28 @@ def test_support_assignment(values, classified, hail_matrix):
     assert anno_matrix.info.categorysupport.collect() == [classified]
 
 
-def test_green_and_new_from_panelapp(panel_changes):
+def test_green_and_new_from_panelapp(mendeliome_expected):
     """
     check that the set expressions from panelapp data are correct
     this is collection of ENSG names from panelapp
     2 set expressions, one for all genes, one for new genes only
-    :param panel_changes:
+    :param mendeliome_expected:
     """
-    green_expression, new_expression = green_and_new_from_panelapp(panel_changes)
+    green_expression, new_expression = green_and_new_from_panelapp(
+        mendeliome_expected['genes']
+    )
 
     # check types
     assert isinstance(green_expression, hl.SetExpression)
     assert isinstance(new_expression, hl.SetExpression)
 
     # check content by collecting
-    assert sorted(list(green_expression.collect()[0])) == [
+    assert sorted(green_expression.collect()[0]) == [
         'ENSG00ABCD',
         'ENSG00EFGH',
         'ENSG00IJKL',
     ]
-    assert list(new_expression.collect()[0]) == ['ENSG00EFGH']
+    assert new_expression.collect()[0] == {'ENSG00IJKL'}
 
 
 @pytest.mark.parametrize(
